fix: file not exits error when SyncToUnity.ps1 run on Windows

+++ DevTools~/SyncToUnity.ps1
@@ -13,5 +13,5 @@ Copy-Item -Path "Resonite~/ResoniteHook/Puppeteer/bin/*.dll" -Destination "./Res
 Copy-Item -Path "Resonite~/ResoniteHook/Puppeteer/bin/*.pdb" -Destination "./ResoPuppet~" -Recurse -Force
 Copy-Item -Path "Resonite~/ResoniteHook/Puppeteer/bin/*.so*" -Destination "./ResoPuppet~" -Recurse -Force
 
-Copy-Item -Path "Resonite~/ResoniteHook/Launcher/bin/Launcher" -Destination "./ResoPuppet~" -Force
-Copy-Item -Path "Resonite~/ResoniteHook/Puppeteer/bin/Puppeteer" -Destination "./ResoPuppet~" -Force
+if (Test-Path "Resonite~/ResoniteHook/Launcher/bin/Launcher"){ Copy-Item -Path "Resonite~/ResoniteHook/Launcher/bin/Launcher" -Destination "./ResoPuppet~" -Force }
+if (Test-Path "Resonite~/ResoniteHook/Puppeteer/bin/Puppeteer"){ Copy-Item -Path "Resonite~/ResoniteHook/Puppeteer/bin/Puppeteer" -Destination "./ResoPuppet~" -Force }
